refactor: move hydration runtime out of declarative module
Rename ensureDeclarativeRuntime to ensureHydrationRuntime and move
to src/hydration/runtime.ts. enableHydration() now imports from
the hydration module directly, with no dependency on declarative/.

declarative/runtime.ts becomes a thin wrapper that calls
ensureHydrationRuntime() + registers declarative debug messages.
