+++ sites/website/src/docs/3.x/resources/export-sizes.md
@@ -19,7 +19,7 @@ Bundle sizes for `@microsoft/fast-element` exports.
 
 | Export | Minified | Gzip | Brotli |
 |--------|----------|------|--------|
-| CDN Rollup Bundle | 65.17 KB | 19.44 KB | 17.34 KB |
+| CDN Rollup Bundle | 64.76 KB | 19.24 KB | 17.19 KB |
 | FASTElement | 23.71 KB | 7.37 KB | 6.63 KB |
 | Updates | 473 B | 337 B | 287 B |
 | Observable | 6.70 KB | 2.49 KB | 2.22 KB |
@@ -34,5 +34,5 @@ Bundle sizes for `@microsoft/fast-element` exports.
 | repeat | 29.57 KB | 9.42 KB | 8.47 KB |
 | css | 2.43 KB | 1.00 KB | 911 B |
 | ElementStyles | 1.65 KB | 729 B | 623 B |
-| enableHydration | 43.33 KB | 13.30 KB | 11.93 KB |
+| enableHydration | 42.93 KB | 13.11 KB | 11.79 KB |
 | ArrayObserver | 12.51 KB | 4.45 KB | 4.01 KB |
